requirements.txt
numpy==1.24.3          # compatible with tensorflow 2.13 on Windows
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.1
scikit-learn==1.3.2
scikit-image==0.22.0
opencv-python==4.8.1.78
tqdm==4.66.1
tensorflow==2.13.0     # GPU build requires CUDA 11.8 + cuDNN 8.6
keras==2.13.1
Pillow==10.1.0
streamlit==1.29.0

